# docs/swagger.yaml
basePath: /api/v1.
definitions:
  APIError:
    properties:
      data:
        $ref: '#/definitions/errorStruct'
      ok:
        type: boolean
    type: object
...
  APIResponse-MessageResponse:
    properties:
      data:
        $ref: '#/definitions/MessageResponse'
      ok:
        type: boolean
    type: object
...
  APIResponse-routes_AuthResponse:
    properties:
      data:
        $ref: '#/definitions/routes.AuthResponse'
      ok:
        type: boolean
    type: object
...
  MessageResponse:
    properties:
      message:
        type: string
    type: object
...
  UserLoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  UserRegisterRequest:
    properties:
      confirmPassword:
        type: string
      email:
        type: string
      name:
        maxLength: 50
        minLength: 3
        type: string
      password:
        maxLength: 50
        minLength: 8
        type: string
      surname:
        maxLength: 50
        minLength: 3
        type: string
      username:
        maxLength: 30
        minLength: 3
        type: string
    required:
    - confirmPassword
    - email
    - name
    - password
    - surname
    - username
    type: object
...
  errorStruct:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  routes.AuthResponse:
    properties:
      accessToken:
        type: string
      expiresAt:
        type: string
    type: object
...
info:
  contact: {}
  title: Mini-Maxit API
  version: 1.0.0
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Logs in a user with email and password, returns JWT tokens
      parameters:
      - description: User Login Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/UserLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/APIResponse-routes_AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIError'
      summary: Login a user
      tags:
      - auth
  /auth/logout:
    post:
      description: Logs out a user by clearing the refresh token cookie
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/APIResponse-MessageResponse'
        "405":
          description: Method Not Allowed
          schema:
            $ref: '#/definitions/APIError'
      summary: Logout a user
      tags:
      - auth
  /auth/refresh:
    post:
      description: Refreshes JWT tokens using a valid refresh token from cookie
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/APIResponse-routes_AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/APIError'
        "405":
          description: Method Not Allowed
          schema:
            $ref: '#/definitions/APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIError'
      summary: Refresh JWT tokens
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Registers a user with name, surname, email, username and password,
        returns JWT tokens
      parameters:
      - description: User Register Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/UserRegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/APIResponse-routes_AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIError'
        "405":
          description: Method Not Allowed
          schema:
            $ref: '#/definitions/APIError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIError'
      summary: Register a user
      tags:
      - auth
...
swagger: "2.0"